Fixed Unexptected mask bits turned ON in QMI Mask
A 32 bit integer LOC_API_ADAPTER_BIT_BATCH_STATUS had
it's 31st bit set to 1 and added into a 64 bit mask, but
since the 31st bit is the sign bit, and is stored as 2s
compliement, this caused the upper 32 bits of the 64 bit
mask to be filled with 1s, which enabled more QMI Mask
bits than expected. This is fixed by assuring bits are
all 64 bit integers.

Bug: 168117915

Test: GPS basic check

CRs-fixed: 2765742
Change-Id: I2806fd3494391d3e729200b4abced1f984d9d401